feat(Auth): Add CredentialFactory for strongly-typed credential creation
Introduces `CredentialFactory` to enable the safe creation of specific `IGoogleCredential` types (e.g., `UserCredential`, `ServiceAccountCredential`) from various sources.
