+++ Src/Support/Google.Apis.Auth/OAuth2/DefaultCredentialProvider.cs
@@ -16,20 +16,13 @@ limitations under the License.
 
 using System;
 using System.IO;
-using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 
-using Google.Apis.Auth.OAuth2.Flows;
-using Google.Apis.Auth.OAuth2.Responses;
-using Google.Apis.Json;
 using Google.Apis.Logging;
-using Google.Apis.Util;
 
 namespace Google.Apis.Auth.OAuth2
 {
-    // TODO(jtattermusch): look into getting rid of DefaultCredentialProvider and moving
-    // the logic into GoogleCredential.
 
     /// <summary>
     /// Provides the Application Default Credential from the environment. 
@@ -156,250 +149,10 @@ namespace Google.Apis.Auth.OAuth2
         {
             Logger.Debug("Loading Credential from file {0}", credentialPath);
 
+            // We use the overridable GetStream method instead of the CredentialFactory.FromFile methods to be able to test ADC.
             using Stream stream = GetStream(credentialPath);
-            return await CreateDefaultCredentialFromStreamAsync(stream, cancellationToken).ConfigureAwait(false);
-        }
-
-        /// <summary>Creates a default credential from a stream that contains JSON credential data.</summary>
-        internal GoogleCredential CreateDefaultCredentialFromStream(Stream stream)
-        {
-            JsonCredentialParameters credentialParameters;
-            try
-            {
-                credentialParameters = NewtonsoftJsonSerializer.Instance.Deserialize<JsonCredentialParameters>(stream);
-            }
-            catch (Exception e)
-            {
-                throw new InvalidOperationException("Error deserializing JSON credential data.", e);
-            }
-            return CreateDefaultCredentialFromParameters(credentialParameters);
-        }
-
-        /// <summary>Creates a default credential from a stream that contains JSON credential data.</summary>
-        internal async Task<GoogleCredential> CreateDefaultCredentialFromStreamAsync(Stream stream, CancellationToken cancellationToken)
-        {
-            JsonCredentialParameters credentialParameters;
-            try
-            {
-                credentialParameters = await NewtonsoftJsonSerializer.Instance
-                    .DeserializeAsync<JsonCredentialParameters>(stream, cancellationToken)
-                    .ConfigureAwait(false);
-            }
-            catch (Exception e)
-            {
-                throw new InvalidOperationException("Error deserializing JSON credential data.", e);
-            }
-            return CreateDefaultCredentialFromParameters(credentialParameters);
-        }
-
-        /// <summary>Creates a default credential from a string that contains JSON credential data.</summary>
-        internal GoogleCredential CreateDefaultCredentialFromJson(string json)
-        {
-            JsonCredentialParameters credentialParameters;
-            try
-            {
-                credentialParameters = NewtonsoftJsonSerializer.Instance.Deserialize<JsonCredentialParameters>(json);
-            }
-            catch (Exception e)
-            {
-                throw new InvalidOperationException("Error deserializing JSON credential data.", e);
-            }
-            return CreateDefaultCredentialFromParameters(credentialParameters);
-        }
-
-
-        /// <summary>Creates a default credential from JSON data.</summary>
-        internal GoogleCredential CreateDefaultCredentialFromParameters(JsonCredentialParameters credentialParameters) =>
-            credentialParameters.ThrowIfNull(nameof(credentialParameters)).Type switch
-            {
-                JsonCredentialParameters.AuthorizedUserCredentialType => new GoogleCredential(CreateUserCredentialFromParameters(credentialParameters)),
-                JsonCredentialParameters.ServiceAccountCredentialType => GoogleCredential.FromServiceAccountCredential(CreateServiceAccountCredentialFromParameters(credentialParameters)),
-                JsonCredentialParameters.ExternalAccountCredentialType => new GoogleCredential(CreateExternalCredentialFromParameters(credentialParameters)),
-                JsonCredentialParameters.ImpersonatedServiceAccountCredentialType => new GoogleCredential(CreateImpersonatedServiceAccountCredentialFromParameters(credentialParameters)),
-                JsonCredentialParameters.ExternalAccountAuthorizedUserCredentialType => new GoogleCredential(CreateExternalAccountAuthorizedUserCredentialFromParemeters(credentialParameters)),
-                _ => throw new InvalidOperationException($"Error creating credential from JSON or JSON parameters. Unrecognized credential type {credentialParameters.Type}."),
-            };
-
-        /// <summary>Creates a user credential from JSON data.</summary>
-        private static UserCredential CreateUserCredentialFromParameters(JsonCredentialParameters credentialParameters)
-        {
-            if (credentialParameters.Type != JsonCredentialParameters.AuthorizedUserCredentialType ||
-                string.IsNullOrEmpty(credentialParameters.ClientId) ||
-                string.IsNullOrEmpty(credentialParameters.ClientSecret))
-            {
-                throw new InvalidOperationException("JSON data does not represent a valid user credential.");
-            }
-            if (credentialParameters.UniverseDomain is not null && credentialParameters.UniverseDomain != GoogleAuthConsts.DefaultUniverseDomain)
-            {
-                throw new InvalidOperationException($"{nameof(UserCredential)} is not supported in universe domains other than {GoogleAuthConsts.DefaultUniverseDomain}");
-            }
-
-            var token = new TokenResponse
-            {
-                RefreshToken = credentialParameters.RefreshToken
-            };
-
-            var initializer = new GoogleAuthorizationCodeFlow.Initializer
-            {
-                ClientSecrets = new ClientSecrets
-                {
-                    ClientId = credentialParameters.ClientId,
-                    ClientSecret = credentialParameters.ClientSecret
-                },
-                ProjectId = credentialParameters.ProjectId
-            };
-            var flow = new GoogleAuthorizationCodeFlow(initializer);
-            return new UserCredential(flow, "ApplicationDefaultCredentials", token, credentialParameters.QuotaProject);
-        }
-
-        /// <summary>Creates a <see cref="ServiceAccountCredential"/> from JSON data.</summary>
-        private static ServiceAccountCredential CreateServiceAccountCredentialFromParameters(
-            JsonCredentialParameters credentialParameters)
-        {
-            if (credentialParameters.Type != JsonCredentialParameters.ServiceAccountCredentialType ||
-                string.IsNullOrEmpty(credentialParameters.ClientEmail) ||
-                string.IsNullOrEmpty(credentialParameters.PrivateKey))
-            {
-                throw new InvalidOperationException("JSON data does not represent a valid service account credential.");
-            }
-            var initializer = new ServiceAccountCredential.Initializer(
-                credentialParameters.ClientEmail, credentialParameters.TokenUri)
-            {
-                ProjectId = credentialParameters.ProjectId,
-                QuotaProject = credentialParameters.QuotaProject,
-                KeyId = credentialParameters.PrivateKeyId,
-                UniverseDomain = credentialParameters.UniverseDomain,
-                UseJwtAccessWithScopes = credentialParameters.UniverseDomain is not null && credentialParameters.UniverseDomain != GoogleAuthConsts.DefaultUniverseDomain
-            };
-            return new ServiceAccountCredential(initializer.FromPrivateKey(credentialParameters.PrivateKey));
-        }
-
-        /// <summary>
-        /// Creates an external account credential from JSON data.
-        /// </summary>
-        private static IGoogleCredential CreateExternalCredentialFromParameters(JsonCredentialParameters parameters)
-        {
-            if (parameters.Type != JsonCredentialParameters.ExternalAccountCredentialType || parameters.CredentialSourceConfig is null)
-            {
-                throw new InvalidOperationException("JSON data does not represent a valid external account credential.");
-            }
-
-            // Build the external credential of the correct type.
-            // The order in which these checks are performed is relevant, see https://google.aip.dev/auth/4117.
-            if (!string.IsNullOrEmpty(parameters.CredentialSourceConfig.EnvironmentId))
-            {
-                if (!parameters.CredentialSourceConfig.EnvironmentId.Equals(AwsExternalAccountCredential.SupportedVersion, StringComparison.OrdinalIgnoreCase))
-                {
-                    throw new InvalidOperationException(
-                        $"Only {AwsExternalAccountCredential.SupportedVersion} version of AWS external account credentials is supported by this library.");
-                }
-
-                return new AwsExternalAccountCredential(new AwsExternalAccountCredential.Initializer(
-                    parameters.TokenUrl, parameters.Audience, parameters.SubjectTokenType,
-                    parameters.CredentialSourceConfig.RegionalCredentialVerificationUrl)
-                {
-                    QuotaProject = parameters.QuotaProject,
-                    ServiceAccountImpersonationUrl = parameters.ServiceAccountImpersonationUrl,
-                    WorkforcePoolUserProject = parameters.WorkforcePoolUserProject,
-                    ClientId = parameters.ClientId,
-                    ClientSecret = parameters.ClientSecret,
-                    RegionUrl = parameters.CredentialSourceConfig.RegionUrl,
-                    SecurityCredentialsUrl = parameters.CredentialSourceConfig.Url,
-                    ImdsV2SessionTokenUrl = parameters.CredentialSourceConfig.ImdsV2SessionTokenUrl,
-                    UniverseDomain = parameters.UniverseDomain,
-                });
-            }
-            else if (!string.IsNullOrEmpty(parameters.CredentialSourceConfig.File))
-            {
-                return new FileSourcedExternalAccountCredential(new FileSourcedExternalAccountCredential.Initializer(
-                    parameters.TokenUrl, parameters.Audience, parameters.SubjectTokenType, parameters.CredentialSourceConfig.File)
-                {
-                    QuotaProject = parameters.QuotaProject,
-                    ServiceAccountImpersonationUrl = parameters.ServiceAccountImpersonationUrl,
-                    WorkforcePoolUserProject = parameters.WorkforcePoolUserProject,
-                    ClientId = parameters.ClientId,
-                    ClientSecret = parameters.ClientSecret,
-                    SubjectTokenJsonFieldName = parameters.ExtractSubjectTokenFieldName(),
-                    UniverseDomain = parameters.UniverseDomain,
-                });
-            }
-            else if (!string.IsNullOrEmpty(parameters.CredentialSourceConfig.Url))
-            {
-                var initializer = new UrlSourcedExternalAccountCredential.Initializer(
-                    parameters.TokenUrl, parameters.Audience, parameters.SubjectTokenType, parameters.CredentialSourceConfig.Url)
-                {
-                    QuotaProject = parameters.QuotaProject,
-                    ServiceAccountImpersonationUrl = parameters.ServiceAccountImpersonationUrl,
-                    WorkforcePoolUserProject = parameters.WorkforcePoolUserProject,
-                    ClientId = parameters.ClientId,
-                    ClientSecret = parameters.ClientSecret,
-                    SubjectTokenJsonFieldName = parameters.ExtractSubjectTokenFieldName(),
-                    UniverseDomain = parameters.UniverseDomain,
-                };
-                if (parameters.CredentialSourceConfig.Headers is object)
-                {
-                    foreach (var header in parameters.CredentialSourceConfig.Headers)
-                    {
-                        initializer.Headers.Add(header);
-                    }
-                }
-
-                return new UrlSourcedExternalAccountCredential(initializer);
-            }
-            else
-            {
-                throw new InvalidOperationException("Unrecognized external credential configuration");
-            }
-        }
-
-        private ImpersonatedCredential CreateImpersonatedServiceAccountCredentialFromParameters(JsonCredentialParameters parameters)
-        {
-            if (parameters.Type != JsonCredentialParameters.ImpersonatedServiceAccountCredentialType
-                || parameters.SourceCredential is null
-                || string.IsNullOrEmpty(parameters.ServiceAccountImpersonationUrl))
-            {
-                throw new InvalidOperationException("JSON data does not represent a valid impersonated service account credential.");
-            }
-
-            // If source credential is of a credential type that does not support impersonation, attemting to create the
-            // impersonated credential will fail a few lines later.
-            var sourceCredential = CreateDefaultCredentialFromParameters(parameters.SourceCredential);
-            var maybeTargetPrincipal = ImpersonatedCredential.ExtractTargetPrincipal(parameters.ServiceAccountImpersonationUrl);
-            var initializer = new ImpersonatedCredential.Initializer(parameters.ServiceAccountImpersonationUrl, maybeTargetPrincipal)
-            {
-                DelegateAccounts = parameters.Delegates?.Length > 0 ? parameters.Delegates.ToList() : null,
-                QuotaProject = parameters.QuotaProject,
-            };
-
-            var impersonatedCredential = ImpersonatedCredential.Create(sourceCredential, initializer);
-            if (parameters.UniverseDomain != parameters.SourceCredential.UniverseDomain)
-            {
-                return (impersonatedCredential as IGoogleCredential).WithUniverseDomain(parameters.UniverseDomain) as ImpersonatedCredential;
-            }
-            return impersonatedCredential;
-        }
-
-        private ExternalAccountAuthorizedUserCredential CreateExternalAccountAuthorizedUserCredentialFromParemeters(JsonCredentialParameters parameters)
-        {
-            if (parameters.Type != JsonCredentialParameters.ExternalAccountAuthorizedUserCredentialType
-                || string.IsNullOrEmpty(parameters.TokenUrl)
-                || string.IsNullOrEmpty(parameters.RefreshToken)
-                || string.IsNullOrEmpty(parameters.ClientId)
-                || string.IsNullOrEmpty(parameters.ClientSecret))
-            {
-                throw new InvalidOperationException("JSON data does not represent a valid external account authorized user credential.");
-            }
-
-            var initializer = new ExternalAccountAuthorizedUserCredential.Initializer(
-                parameters.TokenUrl, parameters.RefreshToken, parameters.ClientId, parameters.ClientSecret)
-            {
-                Audience = parameters.Audience,
-                QuotaProject = parameters.QuotaProject,
-                TokenInfoUrl = parameters.TokenInfoUrl,
-                UniverseDomain = parameters.UniverseDomain,
-            };
-
-            return new ExternalAccountAuthorizedUserCredential(initializer);
+            var credential = await CredentialFactory.FromStreamAsync<IGoogleCredential>(stream, cancellationToken).ConfigureAwait(false);
+            return credential.ToGoogleCredential();
         }
 
         /// <summary>
